Add files via upload
This Is the google data analytics capstone project, it was done with python instead of R, it contains visualizations and interpretations for each chart and code line.
The notebook contains the structure of a data analysis project proposed by the google certificate which is: Ask, Prepare, Process, Analyze, Share and Act Phases. each of this phases are contained in the notebook with their own explained derivables.